Modified main and created parse_bs4.py
